Adding experimental support for AWS-based VMs
Refactored the Vagrantfile to support AWS configuration
information, checked in two additional recipes that are needed for
AWS-based VMs, and fixed a a glitch in the mtsw2e recipe related to
str.each being phased out of Ruby 1.9 (which the AWS-based image uses
whereas the Virtualbox based image is still using 1.8)

The AWS-backed configuration does seem to work, although the Vagrant
plugin needs to mature a bit (e.g. support port forwarding and sync'd
folders better) before most people will want to use it. We'll also want
to add a password to the IPython configuration file to protect notebooks
that are hosted as well...