Enhancements to Exs. 6-3 & 6-13 for Google Takeout
Enhanced Example 6-3 so that it "just works" with Google Takeout data
continaing Unicode where previously an exception was being raised
due to a string being mistakenly decoded twice.

Enhanced Example 6-13 so that email addresses that serve as the basis of
the query are substrings in the From field instead of exact matches,
making it easier to query for known addresses

See also these gists for enhancements with debugging data that may also
be useful.

* https://gist.github.com/ptwobrussell/8791064
* https://gist.github.com/ptwobrussell/8875470